Encapsulate testTree in Tree.main()
