--- src/test/TreeTest.java
@@ -1,9 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-
-/**
- * ADLab02
- * Created by IF on 25.10.17.
- */
-class TreeTest {
-
-}
